refactored code base set up
