+++ NFACT/NFACT_PP/nfactpp_args.py
@@ -1,5 +1,5 @@
 import argparse
-from NFACT.utils.utils import colours
+from NFACT.NFACT_decomp.utils.utils import colours
 
 
 def nfact_pp_args() -> dict:
