+++ pyproject.toml
@@ -30,6 +30,6 @@ requires-python = ">=3.8"
 development = ['ruff']
 
 [project.scripts]
-nfact = "NFACT.__main__:nfact_main"
+nfact = "NFACT.NFACT_decomp.__main__:nfact_main"
 nfact_pp = "NFACT.NFACT_PP.__main__:nfact_pp_main"
 nfact_config = "NFACT.NFACT_config.__main__:nfact_config_main"
